🧬 Pull the common logic of data modals *APIParameter* and *ResponseProperty* up to base class *BaseProperty*.
